requirements.txt
awsebcli==3.12.0
botocore==1.7.33
cement==2.8.2
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
colorama==0.3.7
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
elasticsearch==5.4.0
Flask==0.12.2
Flask-Elasticsearch==0.2.5
idna==2.6
itsdangerous==0.24
Jinja2==2.9.6
jmespath==0.9.3
jsonify==0.5
MarkupSafe==1.0
numpy==1.13.3
pathspec==0.5.0
python-dateutil==2.6.1
PyYAML==3.12
requests==2.9.1
semantic-version==2.5.0
six==1.11.0
tabulate==0.7.5
termcolor==1.1.0
urllib3==1.22
virtualenv==15.1.0
websocket-client==0.44.0
Werkzeug==0.12.2
